requirements.txt
requests==2.31.0
Pillow==10.0.1
pillow_heif==0.13.0
numpy==1.24.3
jsonschema==4.19.0
scipy==1.11.2
exifread==3.0.0